+++ saleor/graphql/app/mutations/app_update.py
@@ -58,7 +58,7 @@ class AppUpdate(DeprecatedModelMutation):
 
         # clean and prepare permissions
         if "permissions" in cleaned_input:
-            permissions = cleaned_input.pop("permissions")
+            permissions = cleaned_input.pop("permissions") or []
             ensure_app_permissions_allowed(permissions)
             cleaned_input["permissions"] = get_permissions(permissions)
             ensure_can_manage_permissions(requestor, permissions)